requirements.txt
tk
customtkinter
Pillow
numpy
opencv-python
imageio
Pmw